Remove unnecessary casts now that *Count is size_t
Bug: None
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/140122
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>

+++ src/dawn/native/d3d/ExternalImageDXGIImpl.cpp
@@ -132,7 +132,7 @@ WGPUTexture ExternalImageDXGIImpl::BeginAccess(
     textureDescriptor.mipLevelCount = mMipLevelCount;
     textureDescriptor.sampleCount = mSampleCount;
     textureDescriptor.viewFormats = mViewFormats.data();
-    textureDescriptor.viewFormatCount = static_cast<uint32_t>(mViewFormats.size());
+    textureDescriptor.viewFormatCount = mViewFormats.size();
 
     DawnTextureInternalUsageDescriptor internalDesc = {};
     if (mUsageInternal != wgpu::TextureUsage::None) {
