+++ src/dawn/native/utils/WGPUHelpers.cpp
@@ -77,7 +77,7 @@ ResultOrError<Ref<BindGroupLayoutBase>> MakeBindGroupLayout(
     }
 
     BindGroupLayoutDescriptor descriptor;
-    descriptor.entryCount = static_cast<uint32_t>(entries.size());
+    descriptor.entryCount = entries.size();
     descriptor.entries = entries.data();
     return device->CreateBindGroupLayout(&descriptor, allowInternalBinding);
 }
