+++ src/dawn/tests/unittests/validation/WritableBufferBindingAliasingValidationTests.cpp
@@ -176,7 +176,7 @@ class WritableBufferBindingAliasingValidationTests : public ValidationTest {
         }
 
         wgpu::BindGroupLayoutDescriptor descriptor;
-        descriptor.entryCount = static_cast<uint32_t>(entries.size());
+        descriptor.entryCount = entries.size();
         descriptor.entries = entries.data();
         return device.CreateBindGroupLayout(&descriptor);
     }
